Support for Elasticsearch 2.x and 5.x
Besides:
- got rid of elasticsearch dependency
- few changes in API
- plugins are now installed using elasticsearch-plugin/plugin tool
- cluster name and tcp transport port are not set by default
