+++ build.gradle
@@ -28,22 +28,20 @@ scmVersion {
 project.version = scmVersion.version
 project.group = 'pl.allegro.tech'
 
-dependencies {
-    compile group: 'org.elasticsearch', name: 'elasticsearch', version: '2.2.0'
-    compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.21'
-    compile group: 'commons-io', name: 'commons-io', version: '2.5'
-    compile group: 'net.lingala.zip4j', name: 'zip4j', version: '1.3.2'
-    compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.4'
+subprojects {
+    apply plugin: 'groovy'
 
-    testCompile group: 'org.codehaus.groovy', name: 'groovy', version: '2.4.6'
-    testCompile group: 'org.spockframework', name: 'spock-core', version: '1.0-groovy-2.4'
-    testCompile group: 'ch.qos.logback', name: 'logback-classic', version: '1.1.7'
-    testCompile group: 'org.skyscreamer', name: 'jsonassert', version: '1.3.0'
-}
+    repositories {
+        mavenCentral()
+    }
 
-test {
-    testLogging {
-        exceptionFormat = 'full'
+    sourceCompatibility = 1.8
+
+    test {
+        testLogging {
+            events "passed", "skipped", "failed"
+            exceptionFormat = 'full'
+        }
     }
 }
