+++ core/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+    compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.21'
+    compile group: 'commons-io', name: 'commons-io', version: '2.5'
+    compile group: 'net.lingala.zip4j', name: 'zip4j', version: '1.3.2'
+    compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.4'
+    compile group: 'org.apache.httpcomponents', name: 'httpclient', version: '4.5.2'
+    compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.6.2'
+    compile group: 'com.fasterxml.jackson.dataformat', name: 'jackson-dataformat-yaml', version: '2.6.2'
+
+    testCompile group: 'org.codehaus.groovy', name: 'groovy', version: '2.4.6'
+    testCompile group: 'org.spockframework', name: 'spock-core', version: '1.0-groovy-2.4'
+    testCompile group: 'ch.qos.logback', name: 'logback-classic', version: '1.1.7'
+    testCompile group: 'org.skyscreamer', name: 'jsonassert', version: '1.3.0'
+}
