+++ core/src/main/java/pl/allegro/tech/embeddedelasticsearch/DocumentWithId.java
