+++ core/src/main/java/pl/allegro/tech/embeddedelasticsearch/EmbeddedElasticsearchStartupException.java
