+++ core/src/main/java/pl/allegro/tech/embeddedelasticsearch/HttpStatusCodes.java
@@ -1,0 +1,6 @@
+package pl.allegro.tech.embeddedelasticsearch;
+
+interface HttpStatusCodes {
+    int OK = 200;
+    int NOT_FOUND = 404;
+}
