+++ core/src/main/java/pl/allegro/tech/embeddedelasticsearch/InvalidSetupException.java
