+++ core/src/main/java/pl/allegro/tech/embeddedelasticsearch/PopularProperties.java
@@ -1,0 +1,7 @@
+package pl.allegro.tech.embeddedelasticsearch;
+
+public interface PopularProperties {
+    String HTTP_PORT = "http.port";
+    String TRANSPORT_TCP_PORT = "transport.tcp.port";
+    String CLUSTER_NAME = "cluster.name";
+}
