+++ core/src/main/java/pl/allegro/tech/embeddedelasticsearch/Require.java
