+++ core/src/main/java/pl/allegro/tech/embeddedelasticsearch/TypeWithMapping.java
@@ -1,0 +1,33 @@
+package pl.allegro.tech.embeddedelasticsearch;
+
+import java.io.IOException;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
+public class TypeWithMapping {
+
+    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+    
+    private final String type;
+    private final JsonNode mapping;
+
+    public TypeWithMapping(String type, String mapping) {
+        this.type = type;
+
+        try {
+            JsonNode json = OBJECT_MAPPER.readTree(mapping);
+            this.mapping = json.get(type);
+        } catch (IOException e) {
+            throw new RuntimeException("Problem with provided mapping for type: " + type, e);
+        }
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public JsonNode getMapping() {
+        return mapping;
+    }
+}
