+++ core/src/test/groovy/pl/allegro/tech/embeddedelasticsearch/ElasticDownloadUrlUtilsSpec.groovy
