+++ es24-test/build.gradle
@@ -1,0 +1,5 @@
+dependencies {
+    testCompile project(':test-base')
+    
+    testCompile group: 'org.elasticsearch', name: 'elasticsearch', version: '2.2.0'
+}
