+++ es50-test/build.gradle
@@ -1,0 +1,8 @@
+dependencies {
+    testCompile project(':test-base')
+
+    testCompile group: 'org.elasticsearch', name: 'elasticsearch', version: '5.0.0'
+    testCompile group: 'org.elasticsearch.client', name: 'transport', version: '5.0.0'
+    testCompile group: 'org.apache.logging.log4j', name: 'log4j-api', version: '2.6.2'
+    testCompile group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.6.2'
+}
