+++ settings.gradle
@@ -1,1 +1,6 @@
 rootProject.name = 'embedded-elasticsearch'
+
+include 'core'
+include 'test-base'
+include 'es24-test'
+include 'es50-test'
