--- src/test/resources/elastic-settings.json
@@ -1,1 +1,0 @@
-{ }
