+++ test-base/build.gradle
@@ -1,0 +1,7 @@
+dependencies {
+    compile project(':core')
+    compile group: 'org.codehaus.groovy', name: 'groovy', version: '2.4.6'
+    compile group: 'org.spockframework', name: 'spock-core', version: '1.0-groovy-2.4'
+    compile group: 'ch.qos.logback', name: 'logback-classic', version: '1.1.7'
+    compile group: 'org.skyscreamer', name: 'jsonassert', version: '1.3.0'
+}
