+++ test-base/src/main/resources/car-mapping.json
