+++ test-base/src/main/resources/paper-book-mapping.json
